+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.3.0",
   "description": "Tailwind plugin for styling scrollbars",
   "author": "Graham Still <graham@gstill.dev>",
-  "main": "index.js",
+  "main": "src/index.js",
   "repository": "https://github.com/adoxography/tailwind-scrollbar",
   "homepage": "https://github.com/adoxography/tailwind-scrollbar",
   "license": "MIT",
@@ -12,7 +12,7 @@
     "test": "jest"
   },
   "files": [
-    "index.js"
+    "src/*"
   ],
   "devDependencies": {
     "jest": "^26.6.3",
